+++ src/apps/SettingsWindow/src/cpp/settings_configuration_snapshot.hpp
@@ -67,6 +67,7 @@ public:
         {"profiles", profiles},
         {"selected_profile",
          {
+             {"ignore_pointing_device_events_by_default", selected_profile.get_ignore_pointing_device_events_by_default()},
              {"parameters",
               {
                   {"delay_milliseconds_before_open_device", selected_profile.get_parameters()->get_delay_milliseconds_before_open_device().count()},