+++ src/share/core_configuration/details/profile.hpp
@@ -31,6 +31,10 @@ public:
                                          selected_,
                                          false);
 
+    helper_values_.push_back_value<bool>("ignore_pointing_device_events_by_default",
+                                         ignore_pointing_device_events_by_default_,
+                                         true);
+
     helper_values_.push_back_object<details::parameters>("parameters",
                                                          parameters_);
 
@@ -38,7 +42,14 @@ public:
                                                                    virtual_hid_keyboard_);
 
     helper_values_.push_back_array<details::device>("devices",
-                                                    devices_);
+                                                    devices_,
+                                                    [this](const auto& json, auto error_handling) {
+                                                      return std::make_shared<details::device>(json,
+                                                                                               error_handling,
+                                                                                               [this](const auto& identifiers) {
+                                                                                                 return make_device_default_values(identifiers);
+                                                                                               });
+                                                    });
 
     //
     // Set default value
@@ -188,6 +199,16 @@ public:
     selected_ = value;
   }
 
+  [[nodiscard]] const bool& get_ignore_pointing_device_events_by_default() const {
+    return ignore_pointing_device_events_by_default_;
+  }
+
+  void set_ignore_pointing_device_events_by_default(bool value) {
+    ignore_pointing_device_events_by_default_ = value;
+
+    update_devices_default_values();
+  }
+
   [[nodiscard]] pqrs::not_null_shared_ptr_t<details::parameters> get_parameters() const {
     return parameters_;
   }
@@ -230,10 +251,14 @@ public:
     // Add device
     //
 
-    devices_.push_back(std::make_shared<details::device>(nlohmann::json({
-                                                             {"identifiers", identifiers},
-                                                         }),
-                                                         error_handling_));
+    auto device = std::make_shared<details::device>(nlohmann::json({
+                                                        {"identifiers", identifiers},
+                                                    }),
+                                                    error_handling_,
+                                                    [this](const auto& identifiers) {
+                                                      return make_device_default_values(identifiers);
+                                                    });
+    devices_.push_back(device);
     return devices_.back();
   }
 
@@ -261,6 +286,7 @@ private:
   error_handling error_handling_;
   std::string name_;
   bool selected_;
+  bool ignore_pointing_device_events_by_default_;
   pqrs::not_null_shared_ptr_t<details::parameters> parameters_;
   pqrs::not_null_shared_ptr_t<simple_modifications> simple_modifications_;
   pqrs::not_null_shared_ptr_t<simple_modifications> fn_function_keys_;
@@ -268,6 +294,38 @@ private:
   pqrs::not_null_shared_ptr_t<details::virtual_hid_keyboard> virtual_hid_keyboard_;
   mutable std::vector<pqrs::not_null_shared_ptr_t<details::device>> devices_;
   configuration_json_helper::helper_values helper_values_;
+
+  void update_devices_default_values() const {
+    for (const auto& device : devices_) {
+      update_device_default_values(*device);
+    }
+  }
+
+  void update_device_default_values(details::device& device) const {
+    device.update_default_values(make_device_default_values(device.get_identifiers()));
+  }
+
+  [[nodiscard]] details::device::default_values make_device_default_values(const device_identifiers& identifiers) const {
+    return {
+        .ignore = make_device_ignore_default_value(identifiers),
+        .manipulate_caps_lock_led = identifiers.get_is_keyboard(),
+    };
+  }
+
+  [[nodiscard]] bool make_device_ignore_default_value(const device_identifiers& identifiers) const {
+    if (identifiers.get_is_game_pad() ||
+        identifiers.get_is_consumer() ||
+        // YubiKey token
+        identifiers.get_vendor_id() == pqrs::hid::vendor_id::value_t(0x1050)) {
+      return true;
+    }
+
+    if (identifiers.get_is_pointing_device()) {
+      return ignore_pointing_device_events_by_default_;
+    }
+
+    return false;
+  }
 };
 
 inline void to_json(nlohmann::json& json, const profile& profile) {